@@ -23,18 +23,21 @@
     "require-dev": {
         "symfony/stopwatch": "~2.8|~3.0",
         "symfony/dependency-injection": "~2.8|~3.0",
-        "symfony/form": "~3.0,>=3.0.5",
+        "symfony/form": "^3.2.5",
         "symfony/http-kernel": "~2.8|~3.0",
         "symfony/property-access": "~2.8|~3.0",
         "symfony/property-info": "~2.8|3.0",
         "symfony/proxy-manager-bridge": "~2.8|~3.0",
         "symfony/security": "~2.8|~3.0",
         "symfony/expression-language": "~2.8|~3.0",
-        "symfony/validator": "~2.8|~3.0",
+        "symfony/validator": "^2.8.18|^3.2.5",
         "symfony/translation": "~2.8|~3.0",
         "doctrine/data-fixtures": "1.0.*",
         "doctrine/dbal": "~2.4",
-        "doctrine/orm": "~2.4,>=2.4.5"
+        "doctrine/orm": "^2.4.5"
+    },
+    "conflict": {
+        "phpunit/phpunit": "<4.8.35|<5.4.3,>=5.0"
     },
     "suggest": {
         "symfony/form": "",
