@@ -17,12 +17,12 @@
     ],
     "require": {
         "php": ">=5.5.9",
-        "twig/twig": "~1.28|~2.0"
+        "twig/twig": "~1.34|~2.4"
     },
     "require-dev": {
         "symfony/asset": "~2.8|~3.0",
         "symfony/finder": "~2.8|~3.0",
-        "symfony/form": "~3.1.9|^3.2.2",
+        "symfony/form": "^3.2.10|^3.3.3",
         "symfony/http-kernel": "~3.2",
         "symfony/polyfill-intl-icu": "~1.0",
         "symfony/routing": "~2.8|~3.0",
@@ -35,6 +35,9 @@
         "symfony/console": "~2.8|~3.0",
         "symfony/var-dumper": "~2.8.10|~3.1.4|~3.2",
         "symfony/expression-language": "~2.8|~3.0"
+    },
+    "conflict": {
+        "symfony/form": "<3.2.10|~3.3,<3.3.3"
     },
     "suggest": {
         "symfony/finder": "",
