@@ -17,6 +17,7 @@
     ],
     "require": {
         "php": ">=5.5.9",
+        "ext-xml": "*",
         "symfony/security": "~3.2",
         "symfony/http-kernel": "~3.2",
         "symfony/polyfill-php70": "~1.0"
@@ -27,20 +28,21 @@
         "symfony/console": "~2.8|~3.0",
         "symfony/css-selector": "~2.8|~3.0",
         "symfony/dom-crawler": "~2.8|~3.0",
-        "symfony/form": "~2.8|~3.0",
-        "symfony/framework-bundle": "~3.1",
+        "symfony/form": "^2.8.18|^3.2.5",
+        "symfony/framework-bundle": "^3.2.5",
         "symfony/http-foundation": "~2.8|~3.0",
         "symfony/security-acl": "~2.8|~3.0",
         "symfony/templating": "~2.8|~3.0",
+        "symfony/translation": "~2.8|~3.0",
         "symfony/twig-bundle": "~2.8|~3.0",
         "symfony/twig-bridge": "~2.8|~3.0",
         "symfony/process": "~2.8|~3.0",
-        "symfony/validator": "~2.8|~3.0",
+        "symfony/validator": "^2.8.18|^3.2.5",
         "symfony/var-dumper": "~3.2",
         "symfony/yaml": "~2.8|~3.0",
         "symfony/expression-language": "~2.8|~3.0",
         "doctrine/doctrine-bundle": "~1.4",
-        "twig/twig": "~1.28|~2.0"
+        "twig/twig": "~1.34|~2.4"
     },
     "suggest": {
         "symfony/security-acl": "For using the ACL functionality of this bundle"
