@@ -27,20 +27,24 @@
         "forum": "https://groups.google.com/forum/#!forum/twig-users"
     },
     "require": {
-        "php": ">=5.2.7"
+        "php": ">=5.3.3"
     },
     "require-dev": {
-        "symfony/phpunit-bridge": "~3.2",
-        "symfony/debug": "~2.7"
+        "symfony/phpunit-bridge": "~3.3@dev",
+        "symfony/debug": "~2.7",
+        "psr/container": "^1.0"
     },
     "autoload": {
         "psr-0" : {
             "Twig_" : "lib/"
+        },
+        "psr-4" : {
+            "Twig\\" : "src/"
         }
     },
     "extra": {
         "branch-alias": {
-            "dev-master": "1.31-dev"
+            "dev-master": "1.34-dev"
         }
     }
 }
