@@ -3,9 +3,10 @@
 namespace TesBoBundle\Controller;
 
 use Symfony\Bundle\FrameworkBundle\Controller\Controller;
-<<<<<<< HEAD
-<<<<<<< HEAD
+use Symfony\Component\Form\Extension\Core\Type\EmailType;
 use Symfony\Component\Form\Extension\Core\Type\FileType;
+use Symfony\Component\Form\Extension\Core\Type\PasswordType;
+use Symfony\Component\Form\Extension\Core\Type\RepeatedType;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\Form\Extension\Core\Type\TextType;
 use Symfony\Component\Form\Extension\Core\Type\DateType;
@@ -13,16 +14,8 @@ use Symfony\Component\Form\Extension\Core\Type\SubmitType;
 use Symfony\Component\Validator\Constraints\DateTime;
 use TesBoBundle\Entity\Medias;
 use TesBoBundle\Entity\MediaTypes;
+use TesBoBundle\Entity\Users;
 
-
-//use AppBundle\Entity\Task;
-=======
-require_once __DIR__ . "vendor/autoload.php";
-use duncan3dc\Speaker\TextToSpeech;
-use duncan3dc\Speaker\Providers\GoogleProvider;
->>>>>>> c2ccd2d5aca7cb553776f323d1aa8ffae0219ff7
-=======
->>>>>>> parent of a6cadbf... added form test add media not working
 
 class DefaultController extends Controller
 {
@@ -45,6 +38,65 @@ class DefaultController extends Controller
         return $medias;
     }
 
+    public function registerAction(Request $request)
+    {
+
+
+
+
+        $em = $this->getDoctrine()->getManager();
+        // just setup a fresh $task object (remove the dummy data)
+        $user = new Users();
+
+        $form = $this->createFormBuilder($user)
+            ->add('pseudo', TextType::class)
+            ->add('password', RepeatedType::class, array(
+                'type' => PasswordType::class,
+                'invalid_message' => 'The password fields must match.',
+                'options' => array('attr' => array('class' => 'password-field')),
+                'required' => true,
+                'first_options'  => array('label' => 'Password'),
+                'second_options' => array('label' => 'Repeat Password')))
+            ->add('email', RepeatedType::class, array(
+                'type' => EmailType::class,
+                'invalid_message' => 'The emails fields must match.',
+                'required' => true,
+                'first_options'  => array('label' => 'email'),
+                'second_options' => array('label' => 'Repeat email')))
+            ->add('save', SubmitType::class, array('label' => 'devenir bo'))
+            ->getForm();
+
+        $form->handleRequest($request);
+
+        if ($form->isSubmitted() && $form->isValid())
+        {
+            $user = $form->getData();
+
+            if ($em->getRepository('TesBoBundle:Users')->findOneByPseudo($form["pseudo"]->getData()))
+                return $this->render('TesBoBundle:Default:register.html.twig', array('form' => $form->createView(),"error"=>"username ".$form["pseudo"]->getData()." already taken"));
+            else if ($em->getRepository('TesBoBundle:Users')->findOneByEmail($form["email"]->getData()))
+                return $this->render('TesBoBundle:Default:register.html.twig', array('form' => $form->createView(),"error"=>"email ".$form["email"]->getData()." already taken"));
+
+            $user->setPassword(hash('sha256', $this->getParameter('salt').$form["password"]->getData()));
+            $user->setRegisteredDatetime(new \DateTime());
+            $user->setIsBanned(false);
+            $em->persist($user);
+            $em->flush();
+
+            return $this->redirectToRoute('tes_bo_homepage');
+        }
+
+
+        return $this->render('TesBoBundle:Default:register.html.twig', array('form' => $form->createView()));
+    }
+
+
+
+
+
+
+
+
     public function indexAction(\Symfony\Component\HttpFoundation\Request $request)
     {
 
@@ -66,11 +118,12 @@ class DefaultController extends Controller
         else {
 
 
-
+/*
             $provider = new GoogleProvider;
 
             $tts = new TextToSpeech("Hello World", $provider);
             $tts->save("/tmp/hello.mp3");
+*/
 
             return $this->render('TesBoBundle:Default:index.html.twig', array('user' => $user, "gallery" => $gallery));
         }
